+++ zenjpeg/src/decode/parser/mod.rs
@@ -15,7 +15,6 @@ mod markers;
 mod output;
 mod progressive;
 mod scan;
-mod transform;
 
 #[cfg(feature = "ultrahdr")]
 use super::extras::MpfImageTypeExt;